Updated: Moved utility classes to the corresponding utility package while also refactoring the classes which uses them.
